Reorganize gas meters into submodules
