+++ crates/module-system/sov-modules-api/src/gas/meters/slot.rs
@@ -1,0 +1,156 @@
+use sov_rollup_interface::da::DaSpec;
+
+use crate::{GasArray, GasMeteringError, Spec, PREFERRED_DATA_FRACTION};
+
+/// A gas meter that tracks the gas used for a slot.
+pub struct SlotGasMeter<S: Spec> {
+    preferred_sequencer: Option<<S::Da as DaSpec>::Address>,
+    initial_slot_gas: S::Gas,
+    // Assumption: The preferred batches/proofs are executed before standard batches/proofs.
+    remaining_preferred_slot_gas: S::Gas,
+    remaining_total_slot_gas: S::Gas,
+}
+
+impl<S: Spec> SlotGasMeter<S> {
+    /// Creates a new `SlotGasMeter`
+    ///
+    /// # Panics
+    /// May panic with an overflow if the PREFFERERD_DATA_FRACTION is defined to be greater than one,
+    /// which is a logic error. Will not panic under normal conditions.
+    pub fn new(
+        remaining_slot_gas: S::Gas,
+        preferred_sequencer: Option<<S::Da as DaSpec>::Address>,
+    ) -> Self {
+        let remaining_preferred_slot_gas = remaining_slot_gas
+            .clone()
+            .scalar_division(PREFERRED_DATA_FRACTION.denominator.into())
+            .checked_scalar_product(PREFERRED_DATA_FRACTION.numerator.into())
+            // This cannot overflow because the PREFERRED_DATA_FRACTION must be less than 1.
+            .unwrap();
+
+        Self {
+            preferred_sequencer,
+            remaining_preferred_slot_gas,
+            initial_slot_gas: remaining_slot_gas.clone(),
+            remaining_total_slot_gas: remaining_slot_gas,
+        }
+    }
+
+    /// Get the remaining slot gas.
+    pub fn remaining_slot_gas(&self, sequencer: &<S::Da as DaSpec>::Address) -> &S::Gas {
+        if Some(sequencer) == self.preferred_sequencer.as_ref() {
+            &self.remaining_preferred_slot_gas
+        } else {
+            &self.remaining_total_slot_gas
+        }
+    }
+
+    /// Get the remaining gas allowed for the preferred sequencer.
+    pub fn remaining_preferred_slot_gas(&self) -> &S::Gas {
+        &self.remaining_preferred_slot_gas
+    }
+
+    /// Charge gas.
+    ///
+    /// # Errors
+    /// Raises an error if the gas to charge is greater than the funds available
+    pub fn charge_gas(
+        &mut self,
+        gas: &S::Gas,
+        sequencer: &<S::Da as DaSpec>::Address,
+    ) -> Result<(), GasMeteringError<S::Gas>> {
+        // Preferred transactions reduce both the "preferred" gas and the "total" gas,
+        // while standard transactions only reduce the `total` gas.
+        // The `initial total` gas is always greater than the "initial preferred" gas, and preferred
+        // transactions are executed before standard transactions.
+        // This mechanism ensures that the preferred sequencer cannot fully censor
+        // other sequencers (or emergency registrations) by exhausting all the available gas.
+        if Some(sequencer) == self.preferred_sequencer.as_ref() {
+            self.remaining_preferred_slot_gas = self
+                .remaining_preferred_slot_gas
+                .checked_sub(gas)
+                .ok_or(self.gas_error(gas.clone(), true))?;
+        }
+
+        self.remaining_total_slot_gas = self
+            .remaining_total_slot_gas
+            .checked_sub(gas)
+            .ok_or(self.gas_error(gas.clone(), false))?;
+
+        Ok(())
+    }
+
+    /// Total gas used in the slot.
+    pub fn total_gas_used(&self) -> S::Gas {
+        self.initial_slot_gas
+            .checked_sub(&self.remaining_total_slot_gas)
+            .expect("The remaining_slot_gas can't be greater than the initial_slot_gas")
+    }
+
+    fn gas_error(&self, gas_to_charge: S::Gas, is_preferred: bool) -> GasMeteringError<S::Gas> {
+        GasMeteringError::SlotOutOfGas {
+            initial_slot_gas: self.initial_slot_gas.clone(),
+            gas_to_charge,
+            remaining_preferred_slot_gas: self.remaining_preferred_slot_gas.clone(),
+            remaining_total_slot_gas: self.remaining_total_slot_gas.clone(),
+            is_preferred,
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use sov_mock_da::{MockAddress, MockDaSpec};
+    use sov_mock_zkvm::MockZkvm;
+    use sov_rollup_interface::execution_mode::Native;
+
+    use crate::{default_spec::DefaultSpec, GasUnit};
+
+    use super::*;
+
+    type S = DefaultSpec<MockDaSpec, MockZkvm, MockZkvm, Native>;
+
+    #[test]
+    fn slot_gas_meter_test() {
+        let mut slot_gas_meter = SlotGasMeter::<S>::new(GasUnit::<2>::from([100, 200]), None);
+        let sequencer = MockAddress::new([10; 32]);
+
+        let gas = GasUnit::<2>::from([10, 20]);
+        slot_gas_meter.charge_gas(&gas, &sequencer).unwrap();
+
+        assert_eq!(
+            slot_gas_meter.remaining_slot_gas(&sequencer),
+            &GasUnit::<2>::from([90, 180])
+        );
+
+        let preferred_sequencer = MockAddress::new([10; 32]);
+        let mut slot_gas_meter =
+            SlotGasMeter::<S>::new(GasUnit::<2>::from([100, 200]), Some(preferred_sequencer));
+
+        let sequencer = MockAddress::new([33; 32]);
+
+        let gas = GasUnit::<2>::from([10, 20]);
+        slot_gas_meter
+            .charge_gas(&gas, &preferred_sequencer)
+            .unwrap();
+
+        let expected_preferred_gas = &GasUnit::<2>::from([80, 160]);
+        assert_eq!(
+            slot_gas_meter.remaining_slot_gas(&preferred_sequencer),
+            expected_preferred_gas
+        );
+
+        let gas = GasUnit::<2>::from([20, 30]);
+        slot_gas_meter.charge_gas(&gas, &sequencer).unwrap();
+
+        assert_eq!(
+            slot_gas_meter.remaining_slot_gas(&sequencer),
+            &GasUnit::<2>::from([70, 150])
+        );
+
+        assert_eq!(
+            &slot_gas_meter.remaining_preferred_slot_gas,
+            expected_preferred_gas
+        );
+    }
+}
