+++ crates/module-system/sov-modules-api/src/gas/meters/unlimited.rs
@@ -1,0 +1,11 @@
+use std::marker::PhantomData;
+
+use crate::{GasMeter, Spec};
+
+/// A [`GasMeter`] that doesn't charge any gas.
+#[derive(Clone, Default)]
+pub struct UnlimitedGasMeter<S>(PhantomData<S>);
+
+impl<S: Spec> GasMeter for UnlimitedGasMeter<S> {
+    type Spec = S;
+}
