+++ crates/module-system/sov-modules-api/src/gas/mod.rs
@@ -1,23 +1,13 @@
 mod metered_utils;
+mod meters;
 mod traits;
 
 #[cfg(test)]
 mod tests;
 
-use std::marker::PhantomData;
-
 pub use metered_utils::{
     charge_gas_to_deserialize_json, metered_credential, MeteredBorshDeserialize,
     MeteredBorshDeserializeError, MeteredHasher, MeteredSigVerificationError, MeteredSignature,
 };
+pub use meters::*;
 pub use traits::*;
-
-use crate::Spec;
-
-/// A [`GasMeter`] that doesn't charge any gas.
-#[derive(Clone, Default)]
-pub struct UnlimitedGasMeter<S>(PhantomData<S>);
-
-impl<S: Spec> GasMeter for UnlimitedGasMeter<S> {
-    type Spec = S;
-}
